Use Gradle 7+
