+++ controlsfx-samples/src/main/resources/META-INF/services/fxsampler.FXSamplerProject
@@ -1,1 +1,1 @@
-org.controlsfx.ControlsFXSampler
+org.controlsfx.ControlsFXSamplerProject
